+++ main.cpp
@@ -7,9 +7,9 @@ int main()
 {
     auto GRID = std::unique_ptr<Container::Grid>(new Container::Grid(2, 2, 1));
 
-    for(int i=0;i<GRID->get_size_x();i++)
+    for(size_t i=0;i<GRID->get_size_x();i++)
     {
-        for(int j=0;j<GRID->get_size_y();j++)
+        for(size_t j=0;j<GRID->get_size_y();j++)
         {
                     auto b1 = GRID->at(i,j,0);
                     b1->insert();
@@ -32,13 +32,13 @@ int main()
     (b1)->insert(o2);
     //(b2)->insert(o1);
 
-    for(int i=0;i<GRID->get_size_x();i++)
+    for(size_t i=0;i<GRID->get_size_x();i++)
     {
-        for(int j=0;j<GRID->get_size_y();j++)
+        for(size_t j=0;j<GRID->get_size_y();j++)
         {
             auto b = GRID->at(i,j,0);
             std::cout << b << "|"  << b->get_self() << " - " ;
-            for(int k=0; b->at(k) != nullptr && k < b->get_obj_size(); k++)
+            for(size_t k=0; b->at(k) != nullptr && k < b->get_obj_size(); k++)
             {
                 auto o = b->at(k);
                 std::cout << o->get_key() << ":" << o->val << ", ";
